backend/pyproject.toml
[project]
name = "judge-training-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=2.6.0",
    "fastapi>=0.122.0",
    "litellm>=1.80.5",
    "opik-optimizer>=2.3.6",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "requests>=2.32.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: marks tests that require running server (deselect with '-m \"not integration\"')",
    "slow: marks tests that make actual LLM calls (deselect with '-m \"not slow\"')",
]
# By default, skip integration tests
addopts = "-m 'not integration'"
